Move comparators and test utils
