+++ collections/_test_utils.ts
@@ -1,0 +1,10 @@
+export class MyMath {
+  multiply(a: number, b: number): number {
+    return a * b;
+  }
+}
+
+export interface Container {
+  id: number;
+  values: number[];
+}
